requirements.txt
setuptools 
wheel
python-dotenv
databricks-sdk[notebook]
nbformat